requirements.txt
# Binance Futures Trading Bot - Python Dependencies

# Core Binance API
python-binance==1.0.19

# Environment variable management
python-dotenv==1.0.0

# HTTP requests
requests==2.31.0

# Data handling
pandas==2.1.4
numpy==1.26.2

# Utilities
certifi==2023.11.17
charset-normalizer==3.3.2
idna==3.6
urllib3==2.1.0

# Date/time utilities
python-dateutil==2.8.2
pytz==2023.3.post1
tzdata==2023.4

# JSON handling
ujson==5.9.0

# Async support (optional, for advanced features)
aiohttp==3.9.1
websockets==12.0